chore: separator fixes and simplification
