+++ src/blocks/box/Box.utils.ts
@@ -1,5 +1,5 @@
 import { getResponsiveCSS } from '../Blocks.utils';
-import { BoxCSSProps, BoxResponsiveCSSPropertiesData, BoxResponsiveProps } from './Box.types';
+import { BoxResponsiveCSSPropertiesData, BoxResponsiveProps } from './Box.types';
 
 const getBoxResponsiveCSSProperties = (props: BoxResponsiveProps): BoxResponsiveCSSPropertiesData[] => [
   { propName: 'align-items', prop: props.alignItems },
@@ -20,25 +20,3 @@ export const getBoxResponsiveCSS = (props: BoxResponsiveProps) => {
   const data = getBoxResponsiveCSSProperties(props);
   return getResponsiveCSS(data);
 };
-
-export const boxCSSPropsKeys: (keyof BoxCSSProps)[] = [
-  'border',
-  'borderRadius',
-  'backgroundColor',
-  'color',
-  'cursor',
-  'position',
-  'boxShadow',
-  'alignItems',
-  'display',
-  'flexDirection',
-  'height',
-  'justifyContent',
-  'margin',
-  'maxHeight',
-  'minHeight',
-  'maxWidth',
-  'minWidth',
-  'padding',
-  'width',
-];
