+++ src/blocks/index.ts
@@ -1,6 +1,6 @@
 export { Box } from './box';
-export { Text } from './text';
 export { Separator } from './separator';
+export { Text } from './text';
 
 export * from './Blocks.colors';
 export * from './Blocks.constants';
