+++ src/blocks/separator/Separator.tsx
@@ -1,10 +1,10 @@
-import { forwardRef } from 'react';
+import React from 'react';
 import styled from 'styled-components';
 
-import { SeparatorCSSProps, SeparatorComponentProps } from './Separator.types';
-import { getSeparatorResponsiveCSS, separatorCSSPropsKeys } from './Separator.utils';
+import { SeparatorCSSProps, SeparatorComponentProps, separatorCSSPropsKeys } from './Separator.types';
+import { getSeparatorResponsiveCSS } from './Separator.utils';
 import { BlockWithoutStyleProp } from 'blocks/Blocks.types';
-import { getBlocksColor } from 'blocks/Blocks.utils';
+import { blocksColors } from 'blocks/Blocks.colors';
 
 export type SeparatorProps = SeparatorCSSProps & SeparatorComponentProps & BlockWithoutStyleProp<HTMLDivElement>;
 
@@ -13,34 +13,28 @@ const StyledSeparator = styled.div.withConfig({
     !separatorCSSPropsKeys.includes(prop as keyof SeparatorCSSProps) && defaultValidatorFn(prop),
 })<SeparatorProps>`
   /* Initial values */
-  width: ${({ width, orientation }) => width || (orientation === 'horizontal' ? 'auto' : '1px')};
-  height: ${({ height, orientation }) => height || (orientation === 'horizontal' ? '1px' : 'auto')};
+  width: ${({ width, orientation }) => width || (orientation === 'horizontal' ? '100%' : '1px')};
+  height: ${({ height, orientation }) => height || (orientation === 'horizontal' ? '1px' : '100%')};
 
   /* Responsive props */
   ${(props) => getSeparatorResponsiveCSS(props)}
 
   /* Non-responsive props */
-  background-color: ${({ backgroundColor }) => getBlocksColor(backgroundColor)};
-  box-shadow: ${({ boxShadow }) => boxShadow};
-  border-radius: ${({ borderRadius }) => borderRadius};
-  border: ${({ border }) => border};
-  position: ${({ position }) => position};
+  background-color: ${blocksColors.lightGray};
 
   /* Extra CSS prop */
   ${({ css }) => css || ''}
 `;
 
-const Separator = forwardRef<HTMLDivElement, SeparatorProps>(
-  ({ backgroundColor = 'lightGray', orientation = 'horizontal', ...props }, ref) => {
-    return (
-      <StyledSeparator
-        ref={ref}
-        {...props}
-        {...{ orientation, backgroundColor }}
-      />
-    );
-  }
-);
+const Separator: React.FC<SeparatorProps> = ({ orientation = 'horizontal', ...rest }) => {
+  return (
+    <StyledSeparator
+      role="separator"
+      {...{ orientation }}
+      {...rest}
+    />
+  );
+};
 
 Separator.displayName = 'Separator';
