+++ src/blocks/separator/Separator.types.ts
@@ -1,37 +1,18 @@
-import { CSSProperties } from 'react';
-
-import { BlocksColors, ResponsiveProp, SpaceType, ValueOf } from '../Blocks.types';
+import { ResponsiveProp, SpaceType, ValueOf } from '../Blocks.types';
 import { FlattenSimpleInterpolation } from 'styled-components';
 
 export type SeparatorResponsiveProps = {
+  /* Sets height css property */
   height?: ResponsiveProp<string>;
   /* Sets margin css property */
   margin?: ResponsiveProp<SpaceType>;
-  /* Sets max-height css property */
-  maxHeight?: ResponsiveProp<string>;
-  /* Sets min-height css property */
-  minHeight?: ResponsiveProp<string>;
-  /* Sets max-width css property */
-  maxWidth?: ResponsiveProp<string>;
-  /* Sets min-width css property */
-  minWidth?: ResponsiveProp<string>;
-  /* Sets padding css property */
-  padding?: ResponsiveProp<SpaceType>;
   /* Sets width css property */
   width?: ResponsiveProp<string>;
 };
 
 export type SeparatorNonResponsiveProps = {
-  /* Sets border css property */
-  border?: string;
-  /* Sets border-radius css property */
-  borderRadius?: string;
-  /* Sets background-color css property */
-  backgroundColor?: BlocksColors;
-  /* Sets position css property */
-  position?: CSSProperties['position'];
-  /* Sets box-shadow css property */
-  boxShadow?: string;
+  /* Orientation of the separator */
+  orientation?: 'horizontal' | 'vertical';
 };
 
 export type SeparatorCSSProps = SeparatorResponsiveProps & SeparatorNonResponsiveProps;
@@ -39,19 +20,9 @@ export type SeparatorCSSProps = SeparatorResponsiveProps & SeparatorNonResponsiv
 export type SeparatorComponentProps = {
   /* Additional prop from styled components to apply custom css to Separator */
   css?: FlattenSimpleInterpolation;
-  /* Orientation of the separator */
-  orientation?: 'horizontal' | 'vertical';
 };
 
-export type SeparatorResponsiveCSSProperties =
-  | 'height'
-  | 'margin'
-  | 'max-height'
-  | 'min-height'
-  | 'max-width'
-  | 'min-width'
-  | 'padding'
-  | 'width';
+export type SeparatorResponsiveCSSProperties = 'height' | 'margin' | 'width';
 
 export type SeparatorResponsivePropValues = ValueOf<SeparatorResponsiveProps>;
 
@@ -59,3 +30,5 @@ export type SeparatorResponsiveCSSPropertiesData = {
   propName: SeparatorResponsiveCSSProperties;
   prop: SeparatorResponsivePropValues;
 };
+
+export const separatorCSSPropsKeys: (keyof SeparatorCSSProps)[] = ['height', 'margin', 'width', 'orientation'];
